pyproject.toml
[project]
name = "mcp-atlassian"
version = "0.2.6"
description = "The Model Context Protocol (MCP) Atlassian integration is an open-source implementation that bridges Atlassian products (Jira and Confluence) with AI language models following Anthropic's MCP specification. This project enables secure, contextual AI interactions with Atlassian tools while maintaining data privacy and security. Key features include:"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "atlassian-python-api>=3.41.16",
    "beautifulsoup4>=4.12.3",
    "httpx>=0.28.0",
    "mcp>=1.3.0",
    "python-dotenv>=1.0.1",
    "markdownify>=0.11.6",
    "markdown>=3.7.0",
    "markdown-to-confluence>=0.3.0",
    "pydantic>=2.10.6",
    "trio>=0.29.0",
    "click>=8.1.7",
    "uvicorn>=0.27.1",
    "starlette>=0.37.1",
]
[[project.authors]]
name = "al"
email = "al@gmail.com"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
al-mcp-atlassian = "mcp_atlassian:main"

[dependency-groups]
dev = [
    "uv>=0.1.0",
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pre-commit>=3.6.0",
    "ruff>=0.3.0",
    "black>=24.2.0",
    "mypy>=1.8.0",
    "mcp[cli]>=1.3.0"
]
